pyproject.toml
[project]
name = "vid2txt"
description = "A Python package for transcribing videos to text using various speech-to-text services"
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.10"

dependencies = [
    "ffmpeg-python>=0.2.0",
    "ipykernel>=6.30.1",
    "jinja2>=3.1.6",
    "requests>=2.32.5",
    "yt-dlp>=2023.12.30",
]

[project.scripts]
vid2txt = "vid2txt.vid2txt_cli:main"

[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "ipython",
    "mypy",
    "pre-commit",
    "pytest",
    "pytest-coverage",
    "ruff",
]

[tool.ruff]
exclude = [".venv/"]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    # Annotations: Enforce type annotations
    "ANN",
    # Complexity: Enforce a maximum cyclomatic complexity
    "C90",
    # Pydocstyle: Enforce docstrings
    "D",
    # Pyflakes.
    "F",
    # Isort: Enforce import order
    "I",
    # Numpy: Enforce numpy style
    "NPY",
    # Print: Forbid print statements
    "T20",
]
ignore = [
    # Allow self and cls to be untyped, and allow Any type
    "ANN101", "ANN102", "ANN401",
    # Pydocstyle ignores
    "D100", "D101", "D104", "D203", "D212", "D401",
    # Allow use of f-strings in logging
    "G004"
]

[tool.pydoclint]
style = "sphinx"
exclude = "test_"
allow-init-docstring = true
arg-type-hints-in-docstring = false
check-return-types = false
require-return-section-when-returning-nothing = false

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
ignore_missing_imports = true

[tool.setuptools]
packages = ["vid2txt"]

[tool.setuptools_scm]
# can be empty if no extra settings are needed, presence enables setuptools_scm

[tool.setuptools.dynamic]
version = {attr = "vid2txt.version.__version__"}
